requirements.txt
# Web framework
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.1

# Database
dj-database-url==2.3.0
psycopg2-binary==2.9.10

# Auth & Security
python-dotenv==1.1.0

# CORS & Static files
django-cors-headers==4.7.0
whitenoise==6.8.2

# Filters & Rate limiting
django-filter==24.3
django-ratelimit==4.1.0
django-redis==5.4.0
redis==5.2.1

# Server
gunicorn==23.0.0

# Payments
mercadopago==2.3.0

# Email
resend==2.7.0

# Utils
Pillow==11.3.0
requests==2.32.3
